feat: StudtItem component 패키지로 분리(#27)
