+++ feature/main/src/main/java/kr/boostcamp_2024/course/main/presentation/MainScreen.kt
@@ -2,27 +2,19 @@ package kr.boostcamp_2024.course.main.presentation
 
 import android.content.res.Configuration
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.MoreVert
 import androidx.compose.material.icons.outlined.Notifications
-import androidx.compose.material.icons.outlined.Star
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.LargeTopAppBar
@@ -39,7 +31,6 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalConfiguration
@@ -50,6 +41,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import kr.boostcamp_2024.course.main.R
+import kr.boostcamp_2024.course.main.component.StudyItem
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -232,79 +224,6 @@ fun StudyTab(
             )
         }
     }
-}
-
-@Composable
-fun StudyItem(
-    studyUrl: String,
-    studyTitle: String,
-    studyDescription: String,
-    studyMember: Int,
-    onStudyClick: () -> Unit,
-) {
-
-    Column {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 16.dp, vertical = 12.dp),
-            horizontalArrangement = Arrangement.spacedBy(16.dp)
-        ) {
-            Icon( // TODO studyUrl
-                modifier = Modifier
-                    .size(56.dp)
-                    .clip(MaterialTheme.shapes.large)
-                    .background(MaterialTheme.colorScheme.outlineVariant),
-                imageVector = Icons.Outlined.Star, contentDescription = null
-            )
-
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .clickable(onClick = onStudyClick)
-            ) {
-                Text(
-                    text = studyTitle,
-                    style = MaterialTheme.typography.bodyLarge
-                )
-
-                if (studyDescription.isNotEmpty()) {
-                    Text(
-                        text = studyDescription,
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-
-                Text(
-                    text = "인원: ${studyMember}명",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant
-                )
-            }
-
-            IconButton(
-                modifier = Modifier.size(24.dp),
-                onClick = { /* TODO : ex) 스터디 나가기 */ }
-            ) {
-                Icon(imageVector = Icons.Default.MoreVert, contentDescription = null)
-            }
-        }
-
-        HorizontalDivider(thickness = 1.dp)
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun StudyItemPreview() {
-    StudyItem(
-        studyUrl = "",
-        studyTitle = "안드로이드 개발자",
-        studyDescription = "안드로이드 개발자를 위한 스터디입니다.",
-        studyMember = 3,
-        onStudyClick = {}
-    )
 }
 
 @Preview(showBackground = true)
